Push `csr` field through pipeline
`csr` field was decoded by `Decoder`, but not present in `Decode`
output layout.

Pushing of this field is necessary to the end of Scheduler. (and
csr number cannot be passed in any other field).

For RS we can either keep the input layouts for RSes and special Units
the same and need to store unnecessary values in generic RS (maybe we can
depend on optimizer) (this commit)
or specifically exclude storing of some fields inside RS (rs out layout would
be specific for standard RS, not used in special units)
or support different layouts at rs selection in scheduler